refactor: reorganize recipes slice and operations
- Move recipes reducer to slice.js
- Fix store configuration to use recipesReducer
- Fix initial state structure
