+++ src/redux/recipes/operations.js
@@ -18,6 +18,9 @@ export const fetchRecipeById = createAsyncThunk(
   async (recipeId, thunkAPI) => {
     try {
       const response = await axiosInstance.get(`/recipes/${recipeId}`);
+      // TODO: remove log
+      console.log(response.data);
+
       return response.data;
     } catch (error) {
       return thunkAPI.rejectWithValue(error.message);
