--- src/redux/recipes/reducer.js
@@ -1,98 +1,0 @@
-import { createSlice } from '@reduxjs/toolkit';
-import {
-  fetchRecipes,
-  fetchRecipeById,
-  addRecipe,
-  updateRecipe,
-  deleteRecipe,
-} from './operations';
-
-const initialState = {
-  items: [],
-  item: null,
-  isLoading: false,
-  error: null,
-};
-
-const recipesSlice = createSlice({
-  name: 'recipes',
-  initialState,
-  reducers: {},
-  extraReducers: builder => {
-    builder
-      // Fetch all recipes
-      .addCase(fetchRecipes.pending, state => {
-        state.isLoading = true;
-        state.error = null;
-      })
-      .addCase(fetchRecipes.fulfilled, (state, action) => {
-        state.isLoading = false;
-        state.items = action.payload;
-      })
-      .addCase(fetchRecipes.rejected, (state, action) => {
-        state.isLoading = false;
-        state.error = action.payload;
-      })
-      // Fetch recipe by ID
-      .addCase(fetchRecipeById.pending, state => {
-        state.isLoading = true;
-        state.error = null;
-      })
-      .addCase(fetchRecipeById.fulfilled, (state, action) => {
-        state.isLoading = false;
-        state.item = action.payload;
-      })
-      .addCase(fetchRecipeById.rejected, (state, action) => {
-        state.isLoading = false;
-        state.error = action.payload;
-      })
-      // Add recipe
-      .addCase(addRecipe.pending, state => {
-        state.isLoading = true;
-        state.error = null;
-      })
-      .addCase(addRecipe.fulfilled, (state, action) => {
-        state.isLoading = false;
-        state.items.push(action.payload);
-      })
-      .addCase(addRecipe.rejected, (state, action) => {
-        state.isLoading = false;
-        state.error = action.payload;
-      })
-      // Update recipe
-      .addCase(updateRecipe.pending, state => {
-        state.isLoading = true;
-        state.error = null;
-      })
-      .addCase(updateRecipe.fulfilled, (state, action) => {
-        state.isLoading = false;
-        const index = state.items.findIndex(
-          item => item._id === action.payload._id
-        );
-        if (index !== -1) {
-          state.items[index] = action.payload;
-        }
-      })
-      .addCase(updateRecipe.rejected, (state, action) => {
-        state.isLoading = false;
-        state.error = action.payload;
-      })
-      // Delete recipe
-      .addCase(deleteRecipe.pending, state => {
-        state.isLoading = true;
-        state.error = null;
-      })
-      .addCase(deleteRecipe.fulfilled, (state, action) => {
-        state.isLoading = false;
-        state.items = state.items.filter(
-          item => item._id !== action.payload
-        );
-      })
-      .addCase(deleteRecipe.rejected, (state, action) => {
-        state.isLoading = false;
-        state.error = action.payload;
-      });
-  },
-});
-
-export const recipesReducer = recipesSlice.reducer;
