+++ src/redux/store.js
@@ -1,10 +1,10 @@
 import { configureStore } from '@reduxjs/toolkit';
 import authReducer from './auth/slice';
-import recipesSlice from './recipes/slice';
+import recipesReducer from './recipes/slice';
 
 export const store = configureStore({
   reducer: {
     auth: authReducer,
-    recipes: recipesSlice,
+    recipes: recipesReducer,
   },
 });
